requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
requests==2.31.0
pandas==2.1.3
matplotlib==3.8.2
pillow==10.1.0
yfinance==0.2.31
python-dotenv==1.0.0
RPi.GPIO==0.7.1
spidev==3.6
waveshare-epd==0.0.0
schedule==1.2.1
